Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main(){
     float pi = 3.14;
-    int r;
-    scanf("%d",&r);
+    float r;
+    scanf("%f",&r);
     float area = (float)pi*r*r;
     printf("Area: %.2f",area);
     return 0;
